refactor: Car, CarManager 파일 domain 디렉토리 하위로 path 변경
